fix: update element for cart
