+++ bookShop/src/app/app.component.html
@@ -3,9 +3,8 @@
   <p>{{descriptionNameShop}}</p>
 </header>
 
-<!-- <app-book></app-book> -->
 <app-books-list></app-books-list>
-<app-cart></app-cart>
+<app-cart-list></app-cart-list>
 
 <footer>
   <p>{{appTitle.textContent}}</p>
