--- bookShop/src/app/cart-list/cart-list.component.html
@@ -1,1 +1,0 @@
-<p>cart-list works!</p>
